Revert interface move
